dag updated

+++ scripts/load_to_remote.sh
@@ -1,14 +1,16 @@
 #!/bin/bash
 
 # Check if the correct number of arguments is provided
-if [ $# -ne 2 ]; then
-  echo "Usage: $0 <i> <project_root>"
+if [ $# -lt 3 ]; then
+  echo "Usage: $0 <version> <project_root> <push_flag>"
+  echo "Where <push_flag> is true to push changes or false to skip."
   exit 1
 fi
 
 # Assign the arguments to variables
-i=$1
+version=$1
 project_root=$2
+push_flag=$3
 
 echo "Saving to DVC and committing to GitHub..."
 echo "Project root directory:"
@@ -56,10 +58,21 @@ fi
 
 $DVC_PATH add "$SAMPLE_CSV_PATH" || { echo "Failed to save data to DVC"; exit 1; }
 $GIT_PATH add "$DVC_FILE_PATH" || { echo "Failed to stage changes for commit"; exit 1; }
-$DVC_PATH push || { echo "Failed to push data to remote DVC repository"; exit 1; }
-$GIT_PATH commit -m "Save validated data number ${i} (testing airflow)"
-$GIT_PATH tag -a "AIRFLOW2.${i}" -m "Version tag" || { echo "Failed to create tag"; exit 1; }
-$GIT_PATH push --tags || { echo "Failed to push tags to remote repository"; exit 1; }
-$GIT_PATH push
-rm "$SAMPLE_CSV_PATH"
-echo "Commit and tag added for file number ${i}."
+$GIT_PATH commit -m "Save validated data with version ${version}"
+$GIT_PATH tag -d "${version}"
+$GIT_PATH tag -a "${version}" -m "Version tag" || { echo "Failed to create tag"; exit 1; }
+
+if [ "$push_flag" == "true" ]; then
+  $GIT_PATH add services/zenml/
+  $GIT_PATH commit -m "Saved zenml db with version ${version}"
+  $GIT_PATH tag -d "${version}"
+  $GIT_PATH tag -a "${version}" -m "Version tag"
+  $DVC_PATH push || { echo "Failed to push data to remote DVC repository"; exit 1; }
+  $GIT_PATH push --tags || { echo "Failed to push tags to remote repository"; exit 1; }
+  $GIT_PATH push
+  rm "$SAMPLE_CSV_PATH"
+else
+  echo "Skipping push due to flag value."
+fi
+
+echo "Commit and tag added for file version ${version}."
